Move user model to models  folder
