+++ routes/web.php
@@ -15,4 +15,4 @@ Route::get('/','StaticPagesController@home')->name('home');
 Route::get('/help','StaticPagesController@help')->name('help');
 Route::get('/about','StaticPagesController@about')->name('about');
 
-Route::get('sigup','UsersController@create')->name('sigup');
+Route::get('/sigup','UsersController@create')->name('sigup');
